% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.GAGA.R
\name{predict_poisson_GAGA}
\alias{predict_poisson_GAGA}
\title{Get predictions from a GAGA poisson model fit object}
\usage{
predict_poisson_GAGA(fit, newx)
}
\arguments{
\item{fit}{Fitted "GAGA" object.}

\item{newx}{Matrix of new values for x at which predictions are to be made. Must be a
matrix.  If the intercept term needs to be considered in the estimation process, then
the first column of \code{X} must be all 1s.}
}
\value{
Predictions
}
\description{
Get predictions from a GAGA poisson model fit object
}
